requirements.txt
spotipy
python-dotenv
pandas
flask  # Añadido para un servidor web si decides servir los datos directamente
